+++ botMain.js
@@ -6,6 +6,7 @@ const commands = requireDir('./Commands');
 const fh = require('./Commands/FileHandler');
 
 const { version } = require('./package.json');
+const adminprefix = fh.get('../Files/local/adminprefix.json');
 const debug = false; //true > no ready-message and BackgroundTasks
 global.bot = new Discord.Client();
 global.guilds = {};
@@ -59,7 +60,10 @@ global.bot.on('message', (message) => { //When Message sent
         }
     }
 
-    if (contentArgs[0].startsWith('uwuadmin')) { //AdminCommands
+    if (contentArgs[0].startsWith(adminprefix)) { //AdminCommands
+
+        message.content = message.content.replace(adminprefix, 'uwuadmin'); //Replacing Prefix for compability
+
         try {
 
             if (!contentArgs[1]) {
@@ -94,9 +98,15 @@ global.bot.login(fh.get('../Files/local/botToken.json').token).catch(err => {
     return;
 }); //Starts Bot
 
+global.bot.on('guildCreate',guild => { //Joining a new Guild while bot is active
+    serverJoin(guild);
+})
+
+
+
 function init() {
 
-    fh.initSettings(); //Init Settings for all Guilds
+    fh.initSettings('all'); //Init Settings for all Guilds
     BotID = '<@!' + global.bot.user.id + '>'; //Get BotID
 
     global.bot.user.setPresence({ game: { name: 'on ' + version }, status: 'online' }); //Set Bot as online
@@ -153,6 +163,10 @@ function init() {
             }
         }
 
+        //Init Admins
+        let Admins = fh.get('../Files/local/' + guild[0] + '/Admins.json');
+        global.guilds[guild[0]]['Admins'] = Admins;
+
         //Ready Message
         if (!debug) {
             StandardChannel.send('I am now ready to use: Prefix \'' + global.guilds[guild[0]]['prefix'] + '\'');
@@ -167,4 +181,57 @@ function init() {
         commands.twitch.checkForStreams();
         commands.Auto.goodbadBot();
     }
+}
+
+function serverJoin(guild) {
+
+    let names = fh.get('../Files/local/names.json');
+    const initset = require('./Files/initsettings.json');
+
+    global.guilds[guild.id] = {}; //Init global.guilds
+
+    //Init settings
+    fh.write('settings.json',initset,guild.id);
+    global.guilds[guild.id]['settings'] = initset; //Init global.guilds.settings         
+
+    //Set StandardChannel
+    for (let ch of guild.channels) {
+
+        if (ch[1].type == 'text') {
+            commands.Channel.set('Standard', ch[1].id, guild.id);
+            StandardChannel = global.bot.channels.get(ch[1].id);
+            break;
+        }
+    }
+
+    //Set Prefix
+    fh.write('prefix.json',"!",guild.id);
+    global.guilds[guild.id]['prefix'] = '!';
+
+    //Add Members to names.json
+    for (let member of guild.members) {
+
+        if (!names[member[1].user.id]) { //Create User
+            names[member[1].user.id] = {};
+        }
+
+        if (!names[member[1].user.id]['guilds']) { //Create guilds Array
+            names[member[1].user.id]['guilds'] = [guild.id];
+
+        } else {
+
+            if (!names[member[1].user.id]['guilds'].includes(guild.id)) { //Add guildID to guilds Array
+                names[member[1].user.id]['guilds'].push(guild.id);
+            }
+        }
+    }
+    fh.write('names.json', names); //Write names back to names.json
+
+    //Init Admins
+    fh.write('Admins.json',[guild.owner.id],guild.id);
+    global.guilds[guild.id]['Admins'] = [guild.owner.id];
+
+    //Ready
+    commands.Channel.get('Standard',guild.id).send('Hello, this is the Standard-Channel of the Server, you can talk to me via !, use !help for help\n I have added '
+    + guild.owner.user.username + ' as Admin on this Server');
 }